Initial Commit
I could probably *maybe* get away with doing this in a more vanilla stack, maybe just TypeScript or something, but I really want to stress myself with learning how to use React, and frameworks in general. I think this project is a good opportunity to try using that with.

The README has a great write-up as to what this project is all about :)