ExifTool Library Catalogging
Figured out an easier way to get all of the metadata about my iTunes music library! It's with a shell script called ExifTool.

```sh
exiftool -r ~/Music/Music/Media/Music/ "-TrackNumber" "-Title" "-Artist" "-Album" "-ContentCreateDate" > ~/Desktop/library.log
```

Lots of helpful links to figure it out, lol

https://unix.stackexchange.com/questions/423955/shell-script-to-read-metadata-of-audio-file
https://formulae.brew.sh/formula/exiftool
https://github.com/exiftool/exiftool
https://exiftool.org/forum/index.php?topic=11781.0
https://exiftool.org/forum/index.php?topic=10595.0
https://exiftool.org/forum/index.php?topic=13251.0
https://stackoverflow.com/questions/13043344/search-and-replace-in-bash-using-regular-expressions (I was going to try figuring out how to format the output logged data using the shell too, but I may do that in the browser/before build-dev time. I might use JSON to store the data, but I'm not totally sure about that yet. Maybe I could gzip the data too? Since it is a lot of JSON. Maybe even use NBT for the memes hehe)
https://exiftool.org/forum/index.php?topic=7783.0

Initially I was trying to do it with ffmpeg, but that wasn't as easy to get the data with. Once I found this tool, it seemed to have a much nicer CLI for this kind of thing I was trying to do. ffmpeg isn't specifically meant for this, it just happens to be able to read EXIF data because it needs it for the file format things.

Eventually I'm going to include more metadata like genres and other artist data and things like that, then I can make it easier to explore the library by other metrics other than just album/artist/song names.

I plan to deal with album art also, I'm not sure if I want to use the images already saved in my library itself, or outsource them to external links that are hosted elsewhere. Either way, I'm going to use a Service Worker for this, so things will continue to work nicely since they will be cached for offline-use, and lighter network response times in general.

I still need to adjust some of the metadata for songs in my library, I think things like genres and dates could be improved a bit.

I also want to track when I discovered the songs/albums individually, I think that's an important part to the music listening journey too. It's part of what helps with reverse-engineering what your sounds are, because you can figure out where you heard certain ideas based on the timeline of when you found them, in comparison to your own music that you worked on which relates to that inspiration.